feat(command): beginning of a command mode
This now gives it the ability to show/hide sources and fields. There
is some rudimentary almost like readline support in it, but it clearly
needs a lot of work. In reality, that readline like support should be
extracted to a new widget in the future.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>