refactor(prompt): refactor command_prompt to a widgets
Rather than wrapping all of the code inside of the UI, it makes more
sense to have the command prompt as an external widget in case I want
to reuse it in another program. It also makes it easier for me to
iterate on the command prompt away from damaging everything.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>