refactor(logstorage): create LogStorage and LogSource
This refactors and cleans up the code a little bit more so there is a
global LogStorage object that is used to store the various logs that
we're working with. Each source has it's own LogSource, which in turn
has the LogEntry records for that source inside of it.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>